seller flow: add groundwork for seller flow modal
This adds a shared FlowModal component that should be used for any multi step
processes or "flows". Renders a full screen modal with added buttons to close
the modal and skip to next or previous steps. To use, pass FlowModal routes as
if to a StackNavgator with a flow order array that defines the order of the
screens to be shown. Flow order defines which component to render based on the
component's displayName property.

For example, for a modal with steps A, B, C, defining a flow modal would look like:
```
class ComponentA extends Component {
  static displayName = "ScreenA"
  ...
}

class ComponentB ...
class ComponentC ...

const modal = FlowModal(
  ["ScreenA", "ScreenB", "ScreenC"],
  {
    "ScreenA": { screen: ComponentA },
    "ScreenB": ...,
    "ScreenC": ...,
  }
)
```

FlowModal is now used in the BecomeSeller flow.