flow modal: refactor to add state to flow component and more flexibility
This makes `flowComponent()` return a component with state. State is important
because navigation should sometimes be able to be disabled. This is relavant to
making the flow modal work for the seller flow's forms.

This also moves the responsibility of rendering the navigation buttons to the
inner component. The inner component is free to navigate around the flow
modal's screens however it wants for more flexibility.