user profile: basic user profile
This adds the skeleton for the user profile page, which can be accessed by
clicking "Profile" (later to be replaced by user's name) in the profile tab.

The user profile page also has a link to the profile editor for editing info
shown on the user profile, and has a link to the become-seller flow.