auth modal: attempt to fix auth modal navigation
This creates new AuthModalLogin and AuthModalRegister components that authmodal should use instead of reusing Login and Register. Although it duplicates some of the functionality, this makes the navigation logic a lot easier.

There are some existing issues with displaying the auth modal as TabNavigator does not seem to unmount component on tab switch. Broken behavior: after opting not to sign in on the auth modal, user is redirected to the home tab. if the unauthed guest tries to nav to the profile tab again, we want the authmodal to show up again. Instead, a blank screen is shown because we returned a 'null' from the `protectedComponent` in charge of navigating to the auth modal.

Merging anyway because we want Paul's checkout to get up to date with some other important changes that this commit brings.