Move react navigation state into redux
This will allow us to customize navigation actions. For example, we can now
easily define behavior like checking if the user is auth'd before switching
to the profile tab and redirecting them to the  auth modal if they are not.
React-navigation's included navigators do not support this kind of navigation
out of the box.

Some of the routes have been commented out for now because we need to update
AuthModal to use the new redux nav state.